+++ deltafi-ui/src/pages/DeltaFileViewerPage.vue
@@ -237,7 +237,7 @@ export default {
         visible: true,
         action: action.name,
         cause: action.errorCause,
-        context: action.errorContext,
+        context: this.formmatContextData(action.errorContext),
       }
     },
     actionRowClass(data) {
@@ -271,6 +271,25 @@ export default {
           });
         }
       });
+    },
+    formmatContextData(contextData) {
+      let formattedString = contextData;
+      // Javascript does not provide the PCRE recursive parameter (?R) which would allow for the matching against nested JSON using regex: /\{(?:[^{}]|(?R))*\}/g. In order to
+      // capture nested JSON we have to have this long regex.
+      let jsonIdentifierRegEx = /\{(?:[^{}]|(\{(?:[^{}]|(\{(?:[^{}]|(\{(?:[^{}]|(\{(?:[^{}]|(\{(?:[^{}]|(\{(?:[^{}]|(\{(?:[^{}]|(""))*\}))*\}))*\}))*\}))*\}))*\}))*\}))*\}/g
+
+      formattedString = formattedString.replace(jsonIdentifierRegEx, match => parseMatch(match));
+      
+      function parseMatch(match) {
+          try {
+            JSON.parse(match);
+          } catch (e) {
+            return match;
+          }
+          return JSON.stringify(JSON.parse(match), null, 2);
+      }
+
+      return formattedString;
     },
   },
   graphQLService: null,