+++ deltafi-ui/src/pages/ErrorsPage.vue
@@ -260,7 +260,7 @@ export default {
       alert(startD + endD);
     },
     openContextDialog(contextData) {
-      this.contextDialogData = contextData;
+      this.contextDialogData = this.formmatContextData(contextData);
       this.showContextDialog = true;
     },
     closeContextDialog() {
@@ -307,6 +307,25 @@ export default {
         this.sortDirection,
         this.ingressFlowNameSelected
       );
+    },
+    formmatContextData(contextData) {
+      let formattedString = contextData;
+      // Javascript does not provide the PCRE recursive parameter (?R) which would allow for the matching against nested JSON using regex: /\{(?:[^{}]|(?R))*\}/g. In order to
+      // capture nested JSON we have to have this long regex.
+      let jsonIdentifierRegEx = /\{(?:[^{}]|(\{(?:[^{}]|(\{(?:[^{}]|(\{(?:[^{}]|(\{(?:[^{}]|(\{(?:[^{}]|(\{(?:[^{}]|(\{(?:[^{}]|(""))*\}))*\}))*\}))*\}))*\}))*\}))*\}))*\}/g
+
+      formattedString = formattedString.replace(jsonIdentifierRegEx, match => parseMatch(match));
+      
+      function parseMatch(match) {
+          try {
+            JSON.parse(match);
+          } catch (e) {
+            return match;
+          }
+          return JSON.stringify(JSON.parse(match), null, 2);
+      }
+
+      return formattedString;
     },
   },
   graphQLService: null,